fix: naive password should use sha1 instead of sha256
